+++ components/search/CardContainer/style.js
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const StyledCardContainer = styled.div`
+  margin: 0 auto;
+  max-width: 67rem;
+`;
