+++ components/search/NoResultMessage/style.js
@@ -1,0 +1,22 @@
+import styled from "styled-components";
+
+export const ImageWrapper = styled.div`
+  margin: auto;
+  text-align: center;
+`;
+
+export const NotFoundTitle = styled.h2`
+  font-family: "Pretendard-Bold";
+  font-size: 1.5rem;
+  color: #343434;
+  text-align: center;
+  margin-bottom: 1.125rem;
+  margin-top: 2.125rem;
+`;
+
+export const NotFoundMessage = styled.h4`
+  font-family: "Pretendard-Medium";
+  font-size: 1.125rem;
+  color: #343434;
+  text-align: center;
+`;
