+++ components/search/SearchTab/index.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+import { TabBox } from "./style";
+
+function SearchTab() {
+  return <TabBox></TabBox>;
+}
+
+export default SearchTab;
