+++ components/search/SearchTab/style.js
@@ -1,13 +1,6 @@
-import React from "react";
 import styled from "styled-components";
 
-function SearchTab() {
-  return <TabBox></TabBox>;
-}
-
-export default SearchTab;
-
-const TabBox = styled.div`
+export const TabBox = styled.div`
   margin-bottom: 3.125rem;
   height: 3.125rem;
   width: 100%;
