+++ pages/search.js
@@ -1,17 +1,14 @@
-import axios from "axios";
 import React, { useState } from "react";
 import Loader from "react-loader-spinner";
 import styled from "styled-components";
 
 import NavBar from "../components/common/NavBar";
 import CardContainer from "../components/search/CardContainer";
-import ListCard from "../components/search/ListCard";
 import NoResultMessage from "../components/search/NoResultMessage";
 import SearchInput from "../components/search/SearchInput";
 import SearchTab from "../components/search/SearchTab";
-import { API_URL } from "../config/index";
 import Layout from "../hocs/Layout";
-function search() {
+function Search() {
   const [searchData, setSearchData] = useState([]);
   const [noData, setNoData] = useState(false);
   const [loading, setLoading] = useState(false);
@@ -38,7 +35,7 @@ function search() {
   );
 }
 
-export default search;
+export default Search;
 
 const LoaderWrapper = styled.div`
   top: 50vh;
